requirements.txt
requests >= 2.31.0
python_dotenv == 1.0.0
python-telegram-bot == 13